+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Change Log
 
 ## 3.0.34
+ - Issue #374 - AL File Wizard - Default settings for API pages:
+   * `alOutline.defaultApiPublisher`: Default Api Publisher value for the page wizard
+   * `alOutline.defaultApiGroup`: Default Api Group value for the page wizard
+   * `alOutline.defaultApiVersion`: Default Api Version value for the page wizard
  - Issue #376 - Add multiple fields: sort field properties alphabetically
 
 Thank you
+ - fvet for reporting issue #374
  - NKarolak for reporting issue #376
 
 ## 3.0.33